requirement.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
Automat==25.4.16
azure-core==1.35.0
azure-storage-blob==12.26.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.20
ci-info==0.3.0
click==8.2.1
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
crewai==0.175.0
crewai-tools==0.0.1
cryptography==45.0.5
cssselect==1.3.0
dataclasses-json==0.6.7
ddgs==9.5.4
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
Faker==37.4.2
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio==5.37.0
gradio_client==1.10.4
groovy==0.1.2
groq==0.30.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
instructor==1.11.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itemadapter==0.12.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.8
litellm==1.74.9
load-dotenv==0.1.0
looseversion==1.3.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
narwhals==2.2.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.102.0
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parsel==1.10.0
parso==0.8.5
passlib==1.7.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
portalocker==2.7.0
posthog==5.4.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.3.2
Protego==0.5.0
protobuf==6.32.0
prov==2.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pyxnat==1.6.3
PyYAML==6.0.2
queuelib==1.8.0
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.3
safehttpx==0.1.6
scikit-learn==1.7.0
scipy==1.16.0
Scrapy==2.13.3
selenium==4.34.2
semantic-version==2.10.0
sentry-sdk==2.32.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
streamlit==1.49.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tldextract==5.3.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.13
uvicorn==0.35.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
